:recycle: move to diary directory
